pyproject.toml
[project]
name = "agentic-classroom"
version = "0.1.0"
description = "Agentic classroom simulation using CrewAI with professor, TA, and student agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.70.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "ai-trace>=0.0.1",
    "xhtml2pdf==0.2.11",
    "markdown>=3.0.0",
    "openai>=1.0.0",
    "setuptools<82",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
